+++ gh-pages-src/pages/dev/Example5.vue
@@ -9,17 +9,7 @@
       <b>Selected:</b> {{ selected || 'not chosen' }}.
       <br><br>
 
-      <form @submit="onSubmit">
-        <cool-select
-          ref="select"
-          v-model="selected"
-          :items="items"
-          placeholder="Select name"
-          @keydown.enter.native="onEnter"
-        />
-      </form>
-
-      <br>
+      {{ selectTextOnFocus }}
 
       <cool-select
         ref="select"
@@ -67,6 +57,7 @@
         :disable-search="disableSearch"
         :select-text-on-focus="selectTextOnFocus"
         placeholder="Select name"
+        simple-input
       >
         <template #input-before>
           before
@@ -148,12 +139,6 @@ export default {
     }
   },
   methods: {
-    onSubmit () {
-      alert('Submit!')
-    },
-    onEnter () {
-      alert('onEnter!')
-    },
     validate () {
       this.errorMessage = this.selected ? null : 'This is required field!'
     }