+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# [3.4.0](https://github.com/iliyaZelenko/vue-cool-select/compare/v3.3.0...v3.4.0) (2019-11-04)
+
+
+### Features
+
+* **props:** simple-input prop ([4c534cc](https://github.com/iliyaZelenko/vue-cool-select/commit/4c534cc))
+
+# [3.3.0](https://github.com/iliyaZelenko/vue-cool-select/compare/v3.2.0...v3.3.0) (2019-10-24)
+
+
+### Features
+
+* **component:** add prop to toggle full input selection on click/focus ([dc1b997](https://github.com/iliyaZelenko/vue-cool-select/commit/dc1b997))
+
 # [3.2.0](https://github.com/iliyaZelenko/vue-cool-select/compare/v3.1.0...v3.2.0) (2019-10-17)